NH056 - Create a separate module for matrix classes
