+++ nhood-engine-core-performance-tests/pom.xml
@@ -32,6 +32,13 @@
         </dependency>
 
         <dependency>
+            <groupId>com.h8.nh</groupId>
+            <artifactId>nhood-engine-matrix</artifactId>
+            <version>0.0.2-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+
+        <dependency>
             <groupId>org.openjdk.jmh</groupId>
             <artifactId>jmh-core</artifactId>
         </dependency>
