+++ nhood-engine-matrix/pom.xml
@@ -4,8 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>nhood-engine-core</artifactId>
-    <version>0.0.2-SNAPSHOT</version>
+    <artifactId>nhood-engine-matrix</artifactId>
 
     <parent>
         <artifactId>nhood-engine</artifactId>
@@ -16,29 +15,16 @@
     <dependencies>
         <dependency>
             <groupId>com.h8.nh</groupId>
+            <artifactId>nhood-engine-matrix-api</artifactId>
+            <version>0.0.2-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>com.h8.nh</groupId>
             <artifactId>nhood-engine-core-api</artifactId>
             <version>0.0.2-SNAPSHOT</version>
         </dependency>
-
-        <dependency>
-            <groupId>com.h8.nh</groupId>
-            <artifactId>nhood-engine-matrix-api</artifactId>
-            <version>0.0.2-SNAPSHOT</version>
-        </dependency>
-
-        <!--Test-->
-        <dependency>
-            <groupId>com.h8.nh</groupId>
-            <artifactId>nhood-engine-core-test</artifactId>
-            <version>0.0.2-SNAPSHOT</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.h8.nh</groupId>
-            <artifactId>nhood-engine-matrix</artifactId>
-            <version>0.0.2-SNAPSHOT</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
-
+    
 </project>
