+++ nhood-engine-matrix/src/main/java/com/h8/nh/nhoodengine/matrix/model/DataMatrixCell.java
@@ -1,4 +1,4 @@
-package com.h8.nh.nhoodengine.core.matrix;
+package com.h8.nh.nhoodengine.matrix.model;
 
 import com.h8.nh.nhoodengine.core.DataResource;
