+++ nhood-engine-matrix/src/main/java/com/h8/nh/nhoodengine/matrix/model/DataMatrixCellConfiguration.java
@@ -1,4 +1,4 @@
-package com.h8.nh.nhoodengine.core.matrix;
+package com.h8.nh.nhoodengine.matrix.model;
 
 import java.math.BigDecimal;
