+++ nhood-engine-matrix/src/test/java/com/h8/nh/nhoodengine/matrix/model/DataMatrixCellConfigurationTest.java
@@ -1,4 +1,4 @@
-package com.h8.nh.nhoodengine.core.matrix;
+package com.h8.nh.nhoodengine.matrix.model;
 
 import org.junit.jupiter.api.Test;
