+++ pom.xml
@@ -17,9 +17,10 @@
 		<module>nhood-engine-core-api</module>
 		<module>nhood-engine-core-performance-tests</module>
 		<module>nhood-engine-core-test</module>
+		<module>nhood-engine-matrix</module>
         <module>nhood-engine-matrix-api</module>
+		<module>nhood-engine-matrix-test</module>
 		<module>nhood-engine-test-utils</module>
-        <module>nhood-engine-matrix-test</module>
     </modules>
 
 	<properties>
